plugin: allow to explicitly ignore any kind of port via udev
The new 'ID_MM_PORT_IGNORE' tag will tell ModemManager to fully avoid using a
given port.

This commit is a the backport of 31f81725e628fd94233e62752ec386215d2b9e13 to
the mm-1-4 branch. Instead of grabbing but ignoring the port, in this case we
just fully ignore it from the very beginning.