mbim: don't assume 'Command Done' is always replied to 'Command' messages
Since libmbim 1.11.1, mbim_device_command_finish() may return messages of type
'Function Error', so we shouldn't assume 'Command Done' messages are returned.

This commit is equivalent to this one in git master, but instead of bumping the
libmbim requirement to 1.11.1, we provide a fallback implementation so that the
same logic works when libmbim is < 1.11.1.

    commit 46b707f8477c2dfbf6e6b4f48c73c2d3cbf5cd12
    Author: Aleksander Morgado <aleksander@aleksander.es>
    Date:   Sun Dec 28 17:55:50 2014 +0100

        mbim: use the new response checker helper method

        mbim_message_response_get_result() is available in libmbim-glib 1.11.1.