broadband-modem-qmi: use 'DMS Set FCC Authentication' if online mode fails
Some new devices, like the Dell DW5770, will return an internal error when
trying to bring the power mode to online. We can avoid this by sending the
magic "DMS Set FCC Auth" message before retrying.

Bumping libqmi version to 1.12.4, which is the one that supports this new
message in the stable libqmi branch.

https://bugzilla.kernel.org/show_bug.cgi?id=92101