proxmark3 is not a modem, blacklist it.
ModemManager locks the proxmark3 when it is connected. As it is not a modem,
ModemManager shouldn't be talking to it.

At present, the recommended action in the proxmark3 documentation is to
purge ModemManager to fix this issue.

Note that this vendor uses an unregistered USB device ID, so it is
instead matched by the manufacturer string.

References:
- http://store.ryscc.com/blogs/news/45802433
- http://www.proxmark.org/forum/viewtopic.php?id=1759